ref(evals): Share eval CLI parsing
Extract common eval script argument parsing and tool-detail environment setup into a shared helper. This keeps workspace and demo eval scripts aligned as the verbosity flags evolve.

Refs #39
